requirements.txt
flask
flask-cors
langchain
langchain-openai
langchain-community
langchain-text-splitters
langchain-chroma
chromadb
openai
tiktoken
python-dotenv
rank-bm25
beautifulsoup4
markdown
pydantic
langchain-experimental==0.0.22
requests
scipy
pandas
numpy